base/requirements.txt
asgiref==3.11.0
bleak==2.1.1
certifi==2026.1.4
charset-normalizer==3.4.4
dbus-fast==3.1.2
Django==6.0.1
idna==3.11
meshtastic==2.7.7
mpmath==1.3.0
packaging==24.2
paho-mqtt==2.1.0
protobuf==6.33.4
pubsub==0.1.2
Pypubsub==4.0.7
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
RPi.GPIO==0.7.1
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tabulate==0.9.0
urllib3==2.6.3
vcgencmd==0.1.1
